Refactoring math methods into separate files.  Instantiating the DP methods for the carrental environment.
